+++ src/main/java/com/team4/domain/comment/service/CommentService.java
@@ -3,10 +3,10 @@ package com.team4.domain.comment.service;
 import com.team4.domain.comment.dto.CommentDto;
 import com.team4.domain.comment.dto.CommentRequestDto;
 import com.team4.domain.comment.entity.Comment;
-import com.team4.domain.member.entity.Member;
+import com.team4.domain.member.dao.MemberRepository;
+import com.team4.domain.member.domain.Member;
 import com.team4.domain.place.entity.Place;
 import com.team4.domain.comment.repository.CommentRepository;
-import com.team4.domain.member.repository.MemberRepository;
 import com.team4.domain.place.repository.PlaceRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -24,14 +24,14 @@ public class CommentService {
     private final PlaceRepository placeRepository;
     private final MemberRepository memberRepository;
 
-    public List<CommentDto> getCommentsByPlace(int placeId) {
+    public List<CommentDto> getCommentsByPlace(Long placeId) {
         List<Comment> comments = commentRepository.findByPlace_PlaceId(placeId);
         return comments.stream()
                 .map(CommentDto::fromEntity)
                 .collect(Collectors.toList());
     }
 
-    public List<CommentDto> getCommentsByPlaceAndIsLocal(int placeId, boolean isLocal) {
+    public List<CommentDto> getCommentsByPlaceAndIsLocal(Long placeId, boolean isLocal) {
         List<Comment> comments = commentRepository.findByPlace_PlaceIdAndIsLocal(placeId, isLocal);
         return comments.stream()
                 .map(CommentDto::fromEntity)
