+++ src/main/java/com/team4/domain/feedback/repository/PlaceFeedbackRepository.java
@@ -3,14 +3,14 @@ package com.team4.domain.feedback.repository;
 
 import com.team4.domain.feedback.entity.FeedbackType;
 import com.team4.domain.feedback.entity.PlaceFeedback;
-import com.team4.domain.member.entity.Member;
+import com.team4.domain.member.domain.Member;
 import com.team4.domain.place.entity.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface PlaceFeedbackRepository extends JpaRepository<PlaceFeedback, Integer> {
+public interface PlaceFeedbackRepository extends JpaRepository<PlaceFeedback, Long> {
 
     Optional<PlaceFeedback> findByPlaceAndMember(Place place, Member member);
